Extracted to test code to superclass
